Remove Google Play Music
The Google Play music integration hasn't been working for a while and
the existing links all redirect to the Youtube Music homepage, so I
don't see any harm in removing it completely. I'll leave the data in the
database just in case it's useful in the future, but it's the first
thing I'd cut if I needed to save some space.