Remove soundcloud and upgrade Rails (#57)
Removing soundcloud since it requires us to pull a browser executable into the image which jacks up the memory usage and costs a lot of money. If we re-implement soundcloud, ideally it would be with an official API.

Upgrading rails prevents an ssl issue we were running into with node by using the new asset pipeline in rails 7. It's not exactly clear to me the tradeoffs of the new asset pipeline, but our assets are extremely vanilla at the moment.